+++ android/quest/src/test/java/org/smartregister/fhirecore/quest/ui/patient/register/PatientMapperTest.kt
@@ -23,9 +23,12 @@ import org.hl7.fhir.r4.model.Patient
 import org.hl7.fhir.r4.model.StringType
 import org.junit.Assert.assertEquals
 import org.junit.Test
+import org.robolectric.annotation.Config
 import org.smartregister.fhircore.quest.ui.patient.register.PatientItemMapper
 import org.smartregister.fhirecore.quest.robolectric.RobolectricTest
+import org.smartregister.fhirecore.quest.shadow.QuestApplicationShadow
 
+@Config(shadows = [QuestApplicationShadow::class])
 class PatientMapperTest : RobolectricTest() {
 
   @Test